refactor: 메세지 팩토리가 메세지를 생성하는 과정을 provider에게 위임
